Refactored the initial starter code into a simulation library.
Elements from the library can be customized to implement a
scheduling policy.

Added a simple random hashing run-to-completion scheduling policy
as an initial exaxmple.

Still need to do more work on the load generator and, of course,
need to add new scheduling policies. Also need to work on the
data analysis/logging aspects of the simulator.